# src/apiserver/openapi_server/openapi/cloudinstance.yaml
openapi: 3.0.4
info:
  title: 5G-META Cloud Platform API Server
  description: "API to manage pipeline instances and instace types in different 5GMETA\
    \ MEC Servers. \nCloud components. The Instance API has the scope to consent the\
    \ request of a \npipeline instance and receive the confirmation of the instance\
    \ deployment. \nAn instance can be deleted, but not be modified or updated.\""
  contact:
    name: 5GMETA
    url: https://5gmeta-project.eu/"
    email: 5gmeta@akkodis.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: /api/v1
...
paths:
...
  /mec:
    get:
      tags:
      - Southbound services
      summary: Get all serving MEC servers
      operationId: get_mecs
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    post:
      tags:
      - Southbound services
      summary: Register a MEC instance in the discovery service
      operationId: post_mec
      requestBody:
        description: MEC information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MECCreation'
        required: true
      responses:
        "200":
          description: Successful operation
          content: {}
        "404":
          description: No services found
          content: {}
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/{mec_id}:
    get:
      tags:
      - Southbound services
      summary: Get a serving MEC server
      operationId: get_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    delete:
      tags:
      - Southbound services
      summary: Delete a MEC
      operationId: delete_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/tile:
    get:
      tags:
      - Southbound services
      summary: Get list of tiles
      operationId: get_tiles
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid tile supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/tile/{tile}:
    get:
      tags:
      - Southbound services
      summary: Find serving MEC in a tile
      operationId: get_mec_tile
      parameters:
      - name: tile
        in: path
        description: Tile to get serving MEC
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "401":
          description: Invalid tile supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
...
  /dataflows/{dataType}/query:
    get:
      tags:
      - Dataflow catalogue
      summary: Returns the list of ids of DataFlows that match the query
      operationId: get_data_flows
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                example:
                - 1
                - 2
                - 3
                - 5
                - 8
                items:
                  type: integer
                x-content-type: application/json
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.dataflow_catalogue_controller
  /dataflows/{dataType}/query/count:
    get:
      tags:
      - Dataflow catalogue
      summary: Returns the number of dataFlows that match the query
      operationId: count_data_flows
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: integer
                example: 5
                x-content-type: text/plain
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.dataflow_catalogue_controller
...
  /topics:
    get:
      tags:
      - Topic management
      summary: Returns the topics associated with that user
      operationId: topics_by_user
      parameters:
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                example:
                - "links_cpm_1, links_cam_1, links_video_1, links_cam_2, links_cam_3"
                items:
                  type: string
                x-content-type: application/json
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.topic_management_controller
  /topics/{topicName}:
    get:
      tags:
      - Topic management
      summary: Returns the query that generated the topic
      operationId: find_query_by_topic_name
      parameters:
      - name: topicName
        in: path
        description: Name of the topic
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.topic_management_controller
    delete:
      tags:
      - Topic management
      summary: Delete a topic for the user
      operationId: delete_topic
      parameters:
      - name: topicName
        in: path
        description: Name of the topic to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content: {}
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.topic_management_controller
  /topics/{dataType}/query:
    post:
      tags:
      - Topic management
      summary: catalogue a Kafka topic for the user and returns the topic
      operationId: create_topic
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: instance_type
        in: query
        description: Instance Type of the pipeline processing the flow
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: links_cpm_1
                x-content-type: text/plain
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.topic_management_controller
...
components:
  schemas:
...
    MECCreation:
      type: object
      properties:
        name:
          type: string
        lat:
          type: string
        lng:
          type: string
        organization:
          type: string
        resources:
          type: object
          properties:
            memory:
              type: string
            cpu:
              type: string
            gpu:
              type: string
            storage:
              type: string
        sb_services:
          type: array
          items:
            $ref: '#/components/schemas/SBService'
        props:
          type: object
        geolocation:
          $ref: '#/components/schemas/GeolocationTileList'
      example:
        name: SanSebastian-MEC
        lat: "43.2922071"
        lng: "-1.987018,17"
        organization: Vicomtech
        resources:
          memory: "16"
          cpu: "12"
          gpu: "true"
          storage: 1000"
        sb_services:
        - service_name: message-broker
          description: Message Broker
          host: "<host>"
          port: "5673"
        - service_name: video-broker
          description: Video Stream Broker
          host: "<host>"
          port: "8443"
        - service_name: registration-api
          description: Registration API
          host: "<host>"
          port: "12346"
        props: {}
        geolocation:
        - tile-id: "0331110121011323"
        - tile-id: "0331110121023020"
      xml:
        name: MECInstance
...
    MECInstanceList:
      type: array
      items:
        $ref: '#/components/schemas/MECInstance'
...
    GeolocationTileList:
      type: array
      items:
        $ref: '#/components/schemas/GeolocationTile'
...